app/requirements.txt
selenium==4.23.1
schedule==1.2.0
python-dotenv==1.0.1
